refactor: remove module inception and make clippy happy
