+++ tycho-indexer/src/extractor/evm/token_analysis_cron.rs
@@ -14,8 +14,7 @@ use tycho_core::{
     Bytes,
 };
 use tycho_ethereum::{
-    token_analyzer::trace_call::TraceCallDetector,
-    token_pre_processor::token_pre_processor::map_vault,
+    token_analyzer::trace_call::TraceCallDetector, token_pre_processor::map_vault,
 };
 
 pub async fn analyze_tokens(
